+++ daemon/network/message_handler/add.rs
@@ -11,7 +11,12 @@ use crate::state_helper::save_state;
 /// Invoked when calling `pueue add`.
 /// Queues a new task to the state.
 /// If the start_immediately flag is set, send a StartMessage to the task handler.
-pub fn add_task(message: AddMessage, sender: &Sender<Message>, state: &SharedState) -> Message {
+pub fn add_task(
+    message: AddMessage,
+    sender: &Sender<Message>,
+    state: &SharedState,
+    settings: &Settings,
+) -> Message {
     let mut state = state.lock().unwrap();
     if let Err(message) = ensure_group_exists(&mut state, &message.group) {
         return message;
@@ -61,7 +66,7 @@ pub fn add_task(message: AddMessage, sender: &Sender<Message>, state: &SharedSta
 
     // Add the task and persist the state.
     let task_id = state.add_task(task);
-    ok_or_return_failure_message!(save_state(&state));
+    ok_or_return_failure_message!(save_state(&state, settings));
 
     // Notify the task handler, in case the client wants to start the task immediately.
     if message.start_immediately {