+++ daemon/network/message_handler/log.rs
@@ -2,11 +2,12 @@ use std::collections::BTreeMap;
 
 use pueue_lib::log::read_and_compress_log_file;
 use pueue_lib::network::message::*;
+use pueue_lib::settings::Settings;
 use pueue_lib::state::SharedState;
 
 /// Invoked when calling `pueue log`.
 /// Return tasks and their output to the client.
-pub fn get_log(message: LogRequestMessage, state: &SharedState) -> Message {
+pub fn get_log(message: LogRequestMessage, state: &SharedState, settings: &Settings) -> Message {
     let state = { state.lock().unwrap().clone() };
     // Return all logs, if no specific task id is specified.
     let task_ids = if message.task_ids.is_empty() {
@@ -24,7 +25,7 @@ pub fn get_log(message: LogRequestMessage, state: &SharedState) -> Message {
             let output = if message.send_logs {
                 match read_and_compress_log_file(
                     *task_id,
-                    &state.settings.shared.pueue_directory(),
+                    &settings.shared.pueue_directory(),
                     message.lines,
                 ) {
                     Ok(output) => (Some(output)),