+++ lib/tests/settings_backward_compatibility.rs
@@ -10,7 +10,7 @@ use pueue_lib::settings::Settings;
 ///
 /// We have to be able to restore from that config at all costs.
 /// Everything else results in a breaking change and needs a major version change.
-/// (For `pueue_lib` as well as `pueue`!
+/// (For `pueue_lib` as well as `pueue`!)
 ///
 /// On top of simply having old settings, I also removed a few default fields.
 /// This should be handled as well.